requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
environ==1.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
